Removed GLFW dependency from Clock class

+++ Spark/Clock.h
@@ -12,7 +12,7 @@ class Clock
     public:
     static void tick();
     static double getDeltaTime();
-    static int getFPS();
+    static double getFPS();
 };
 
 }  // namespace spark
